Initial import
Initial import of all artwork and code. The code is probably going to be
completely ripped out after I refactor currently to use pre-rendered
bitmap components for speed reasons (vector graphics are too laggy to
construct on-the-fly, unfortunately). Also makes for cleaner code and
nicer visuals (nice airplane in the center of the gyrocompass).