COR problems fixed and introduced POIs
Center of rotation of CDI needle was indeed because needle is not
centered (arrowtip side is a bit longer). Added manual COR points and
now it works perfectly. Also introduced point-of-interest markers in
order to avoid having hardcoded information (such as
pixels-per-degree-pitch). Simply put the markers on the appropriate
points and the program logic calculates everything from it.