Replaced stencil buffer by Scissor Test
Scissor test is much faster than messing around with the stencil buffer,
which is quite slow on the Raspberry Pi (~20ms) and therefore ruins
performance. However, requires window coordinates and therefore depends
on the window size, which is currently hardcoded. I.e., does only work
in 720 window mode for now.